fix bug with type

+++ tasks/maslova_u_char_frequency_count/mpi/src/ops_mpi.cpp
@@ -57,11 +57,11 @@ bool MaslovaUCharFrequencyCountMPI::RunImpl() {
   if (rank == 0) {
     size_t part = input_str_size / proc_size;
     size_t rem = input_str_size % proc_size;
-    for (int i = 0; i < proc_size; ++i) {
+    for (size_t i = 0; i < proc_size; ++i) {
       send_counts[i] = part + (i < rem ? 1 : 0); //общий размер, включающий остаток, если он входит
     }
     displs[0] = 0;
-    for (int i = 1; i < proc_size; ++i) {
+    for (size_t i = 1; i < proc_size; ++i) {
       displs[i] = displs[i-1] + send_counts[i-1];
     }
   }
